[vnode] Make vnode versioning more robust.
This patch changes version read/write path such that it does not
allocate heap memory while dealing with versioning. Instead, because
version data is short, it uses stack allcocated array to hold the versioning
data. This eliminate one potential failure in read/write version, which is
used to guarantee data consistency.

* Bonus fix
The patch also fixes a bunch of failure handling in the path of vnode
write.