[vnode] Handle failure of adding to hash in vnodeOpen().
This patch adds error handling for call to vnodeOpen() in vnodeOpen.
Previously the error is not handled, which could cause the allocated
pVnode not registered anywhere, and this could cause memory leak.

This patch also adds a missing deallocation of pVnode on failure.
taosDeleteIntHash() only removes the pVnode from hash table, but
pVnode needs to be deallocated explicitly.