fix(legacy): use describe to build sql when 0x2600
Currently, `show create table` result not always be correct.
It will be truncated when the sql length is large than 65535.

The truncated sql will cause error 0x2600 (syntax error).

Then we'll use describe result to build the whole sql.
Note that this method does not guarantee that the source and
target schema is the same when using with TTL and other table-specified
options. It just works for common usage.

Close [TS-4323](https://jira.taosdata.com:18080/browse/TS-4323)